@@ -1,16 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:slivers_app/screens/sliver_screen.dart';
+
+import 'screens/sliver_screen.dart';
 
 void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
-        debugShowCheckedModeBanner: false,
-        title: 'Material App',
-        home: SliverScreen());
+      debugShowCheckedModeBanner: false,
+      title: 'Material App',
+      home: SliverScreen(),
+    );
   }
 }
